Fix lists with unwrapped text children
The text editor normalises list items to always contain a paragraph, but this was done post migration so some legacy licences contain un-normalised list data.

Where likely occurrences of this are found - specifically where `renderNode` is called on a text node - then render the text node wrapped in a paragraph.