Whitelist fields for comment duplication backwards compatability fix
There was a fix to incude comments made on unprefixed field names for backwards compat with a bugfix for legacy animals section (https://github.com/UKHomeOffice/asl-projects/pull/376)

However, this now results in comments being duplicated for all fields which share a name in a repeated section and at the top level (e.g. `title`).

Fix this by whitelisting only to the original set of field names that was affected - namely those in a legacy PPL animals section.