Cast all rowSpan/colSpan attributes to integers
Occasionally they come back as strings, which means that when they are added together to get the total number of columns in a table then the result is `11111`, which breaks pretty badly for obvious reasons.

Apply a `parseInt` every time we get a `colSpan`/`rowSpan` attribute to make sure that summing the values makes normal mathematical sense.