Separate experience fields from full schema
These are used in PPL transfers to read the field config, but by loading the full schema it includes basically the entire projects app because all the page components are imported.

Pulling the experience fields out into a separate module and importing that directly reduces the bundle size for the task page by ~2-3mb.