Fix image rendering when nested inside other nodes
When rendering an image inside another node the renderer failed because `renderNode` pretends it has the document as the parent node. Except `addImage` doesn't exist on the parent node when it's _not_ the document, so the renderer throws with an error.

Add the document as a `const` in the top level scope rather than passing "probably the document" around as an argument so that it can be reliably referenced.