Use comparison of local and saved states to check synced-ness
Instead of keeping track of the flag that determines if a sync is _currently_ in progress, use the difference between local and saved states to determine if the data has been saved.

This ensures that any local changes will always result in alerting the user, even if we're in the debounce period between syncs.