Only show one comparison tab when previous and granted versions are the same
In the case where an amendment is on it's first iteration we only want to show the granted lcience in the comparison window, and not add a second tab for "previous version" which shows the same version.

Check that the previous version is not granted and only show a second tab where there are enough versions to warrant the extra level of comparison.