Fix saving legacy conditions
Since the addition of reminders the edit condition component now emits an object as a value rather than a string so need to save the `content` property of the emitted value otherwise the condition is rendered as `[object Object]`.

Also fix a weird bug where if an inspector visits the introductory details page before editing conditions then the syncing indicator never completes because it thinks there's an unsynced change to the `holder` property on the local values.