Prevent empty duration field putting application into permanent syncing state
When the duration field is empty it is set to `{ months: undefined, years: undefined }`, where the saved state is `{}`.

This causes `isEqual` to detect a difference, but the `diff` function does not create a patch as it considers the objects to be the same.

That means that `isSyncing` is set to true, but a sync is never carried out, and so the application gets stuck in that state until such a time that the duration field is set.

If no patch can be generated between the saved and local state then assume they are equivalent and dispatch `doneSyncing` to allow subsequent sync actions to be carried out, and set the saved state to the local state.