Default checksum omission to `id`
The version data is extended on the client with the id, which means it will always differ from the server value, which does not have an id value.

This prevents an error being thrown in RAs where no omissions are defined on the server side because an RA does not include any server computed properties.

Also improve error handling so that any client side errors in the sync pipeline are correctly logged. Currently if there is an error thrown during the sync pipeline (such as the above) it is not reported.